do: Finalizando aplicativo. Adicionado funcionalidade da tela de Livros, adicionado tela de Perfil com possibilidade de editar as opções do perfil, e ao mesmo tempo podendo sair ou deletar do mesmo.

+++ app/(tabs)/user-analises.tsx
@@ -29,7 +29,10 @@ export default function UserAnalises() {
 
 	const handleDelete = (id: string) => {
 		if (Platform.OS === "web") {
-			PHContentHandler.handleDeleteUserAnalise(id, setMyAnalises);
+			if (window.confirm("Deseja mesmo apagar sua publicação?")) {
+				PHContentHandler.handleDeleteUserAnalise(id, setMyAnalises);
+			}
+
 			return;
 		}
